Fixed keyboard from invite link, fixed Show group members
