+++ app/handlers/group_leader.py
@@ -5,7 +5,7 @@ from aiogram.types import Message
 from aiogram.fsm.state import StatesGroup, State
 from aiogram.fsm.context import FSMContext
 from app.db.repository import GroupRepo, UserRepo
-from app.keyboards.reply import get_main_menu_leader
+from app.keyboards.reply import get_main_menu_leader, get_regular_member_menu
 from datetime import datetime, timedelta
 import uuid  # Для генерации уникальных идентификаторов
 
@@ -78,42 +78,14 @@ async def process_invite_link(message: Message, state: FSMContext, user_repo: Us
 
         await group_repo.add_member(group_id=group.id, user_id=user.telegram_id, is_leader=False)
         await state.clear()
-        await message.answer(f"Вы успешно присоединились к группе «{group.name}»!")
+        await message.answer(
+            f"Вы успешно присоединились к группе «{group.name}»!",
+            reply_markup=get_regular_member_menu())
     except Exception as e:
         logger.error(f"Ошибка в process_invite_link: {e}")
         await state.clear()
         await message.answer("Произошла ошибка при присоединении. Попробуйте позже.")
         
-@router.message(F.text == "👥 Участники группы")
-async def handle_group_members_leader(message: Message, user_repo: UserRepo, group_repo: GroupRepo):
-    """Обработчик для старосты: отображение списка участников группы с ролями."""
-    try:
-        user = await user_repo.get_user_with_group_info(message.from_user.id)
-        if not user or not user.group_membership or not user.group_membership.is_leader:
-            await message.answer("У вас нет прав для просмотра участников группы.")
-            return
-
-        group = user.group_membership.group
-        members = await group_repo.get_group_members(group.id)
-        if not members:
-            await message.answer("В группе пока нет участников.")
-            return
-
-        # Формируем список участников
-        member_list = []
-        for member in members:
-            member_user = await user_repo.get_user_with_group_info(member.user_id)
-            if member_user:
-                role = "Староста" if member.is_leader else "Ассистент" if member.is_assistant else "Участник"
-                member_info = f"{member_user.first_name} {member_user.last_name or ''} (@{member_user.telegram_username or 'без имени'}) - {role}"
-                member_list.append(member_info)
-
-        response = f"Участники группы «{group.name}»:\n" + "\n".join(member_list)
-        await message.answer(response)
-    except Exception as e:
-        logger.error(f"Ошибка в handle_group_members: {e}", exc_info=True)
-        await message.answer("Произошла ошибка при получении списка участников. Попробуйте позже.")
-
 @router.message(F.text == "📅 События и Бронь")
 async def handle_events_and_booking(message: Message, group_repo: GroupRepo, user_repo: UserRepo):
     try:
