+++ app/handlers/regular_member.py
@@ -24,6 +24,7 @@ async def show_menu(message: Message, user_repo: UserRepo):
         logger.error(f"Ошибка в show_menu: {e}")
         await message.answer("Произошла ошибка. Попробуйте позже.")
 
+'''
 @router.message(Command("member_info"))
 @router.message(F.text == "ℹ️ Информация о группе")
 async def show_member_info(message: Message, user_repo: UserRepo):
@@ -37,6 +38,7 @@ async def show_member_info(message: Message, user_repo: UserRepo):
     except Exception as e:
         logger.error(f"Ошибка в show_member_info: {e}")
         await message.answer("Произошла ошибка. Попробуйте позже.")
+'''
 
 @router.message(Command("calendar"))
 @router.message(F.text == "📅 Показать календарь")
@@ -60,27 +62,32 @@ async def show_weekly_calendar_member(message: Message, user_repo: UserRepo, gro
         logger.error(f"Ошибка в show_weekly_calendar_member: {e}")
         await message.answer("Произошла ошибка. Попробуйте позже.")
 
-@router.message(Command("group_members"))
 @router.message(F.text == "👥 Участники группы")
-async def show_group_members(message: Message, user_repo: UserRepo, group_repo: GroupRepo):
-    """Обработчик для отображения списка участников группы."""
+async def handle_group_members_leader(message: Message, user_repo: UserRepo, group_repo: GroupRepo):
+    """Обработчик: отображение списка участников группы с ролями."""
     try:
         user = await user_repo.get_user_with_group_info(message.from_user.id)
         if not user or not user.group_membership:
-            await message.answer("Вы не состоите в группе.")
+            await message.answer("У вас нет прав для просмотра участников группы.")
             return
 
         group = user.group_membership.group
         members = await group_repo.get_group_members(group.id)
         if not members:
-            await message.answer("В группе нет участников.")
+            await message.answer("В группе пока нет участников.")
             return
 
-        member_list = "\n".join(
-            f"- {m.user.first_name} {m.user.last_name or ''} (@{m.user.telegram_username or 'нет имени'})"
-            for m in members
-        )
-        await message.answer(f"Участники группы «{group.name}»:\n{member_list}")
+        # Формируем список участников
+        member_list = []
+        for member in members:
+            member_user = await user_repo.get_user_with_group_info(member.user_id)
+            if member_user:
+                role = "Староста" if member.is_leader else "Ассистент" if member.is_assistant else "Участник"
+                member_info = f"{member_user.first_name} {member_user.last_name or ''} (@{member_user.telegram_username or 'без имени'}) - {role}"
+                member_list.append(member_info)
+
+        response = f"Участники группы «{group.name}»:\n" + "\n".join(member_list)
+        await message.answer(response)
     except Exception as e:
-        logger.error(f"Ошибка в show_group_members: {e}")
-        await message.answer("Произошла ошибка. Попробуйте позже.")
+        logger.error(f"Ошибка в handle_group_members: {e}", exc_info=True)
+        await message.answer("Произошла ошибка при получении списка участников. Попробуйте позже.")
