+++ app/keyboards/reply.py
@@ -42,7 +42,6 @@ def get_regular_member_menu() -> ReplyKeyboardMarkup:
     return ReplyKeyboardMarkup(
         keyboard=[
             [KeyboardButton(text="👥 Участники группы")],
-            [KeyboardButton(text="ℹ️ Информация о группе")],
             [KeyboardButton(text="📅 Показать календарь")],
             [KeyboardButton(text="📅 Показать недельный календарь")]
         ],
